requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
asyncpg
pydantic-settings
python-multipart
pillow # for convertation